bot/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.27.2
python-dotenv==1.0.1
gspread==6.1.4
google-auth==2.35.0
websockets==13.1
pydantic==2.9.2
